Make rate limit working again
