+++ malwarecage/app.py
@@ -1,18 +1,18 @@
-import time
-
 from datetime import datetime
 
 import redis
 
 from flask import request, g
 from flask_migrate import Migrate
-from werkzeug.exceptions import Forbidden, TooManyRequests
+from werkzeug.exceptions import Forbidden
+from werkzeug.routing import BaseConverter
 
 from malwarecage.core.app import app, api
 from malwarecage.core.capabilities import Capabilities
 from malwarecage.core.config import app_config
 from malwarecage.core.log import getLogger
 from malwarecage.core.plugins import PluginAppContext, load_plugins
+from malwarecage.core.rate_limit import rate_limit
 from malwarecage.core.static import static_blueprint
 from malwarecage.core.util import token_hex
 from malwarecage.model import db, User, APIKey
@@ -47,6 +47,28 @@ from malwarecage.resources.user import (
 )
 
 
+class HashConverter(BaseConverter):
+    # MD5/SHA1/SHA256/SHA512 (32,40,64,128)
+    regex = '(root|[A-Fa-f0-9]{32,128})'
+
+
+app.config["SQLALCHEMY_DATABASE_URI"] = app_config.malwarecage.postgres_uri
+app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
+app.config["SECRET_KEY"] = app_config.malwarecage.secret_key
+"""
+Flask-restful tries to be smart and transforms NotFound exceptions.
+Adds: "You have requested this URI [/file/root] but did you mean
+/file/<string:identifier> or /file ?" even if URI is completely correct,
+but some objects needed to fulfill request are missing.
+"""
+app.config["ERROR_404_HELP"] = False
+
+# Load Flask-specific overrides if specified
+if app_config.malwarecage.flask_config_file:
+    app.config.from_pyfile(app_config.malwarecage.flask_config_file)
+
+app.url_map.converters['hash64'] = HashConverter
+
 migrate = Migrate(app, db, directory=migrations_dir)
 db.init_app(app)
 
@@ -75,24 +97,7 @@ def log_request(response):
     return response
 
 
-def rate_limit(conn, key, duration, limit):
-    class TooManyRequestsWithRetryAfter(TooManyRequests):
-        def __init__(self, retry_after):
-            self.retry_after = retry_after
-            super().__init__("You are too fast. Wait {} seconds before next request.".format(retry_after))
-
-        def get_headers(self, environ=None):
-            return super().get_headers(environ=environ) + [("Retry-After", self.retry_after)]
-
-    current_time = time.time()
-    rate_key = "{}-{}:{}:{}".format(key, g.auth_user.login, duration, current_time // duration)
-    count = conn.incr(rate_key)
-    conn.expire(rate_key, duration)
-    if count > limit:
-        raise TooManyRequestsWithRetryAfter(duration - int(current_time % duration))
-
-
-@app.before_request
+@api.before_request
 def require_auth():
     if request.method == 'OPTIONS':
         return
@@ -124,29 +129,28 @@ def require_auth():
             """
             Single sample view in malwarefront generates 7 requests (6 GET, 1 POST)
             """
-            conn = redis.from_url(app_config.malwarecage.redis_uri)
             if request.method == 'GET':
                 """
                 DownloadResource is token-based and shouldn't be limited
                 """
-                if request.endpoint != 'downloadresource':
+                if request.endpoint != api.endpoint_for(DownloadResource):
                     # 1000 per 10 seconds
-                    rate_limit(conn, "get-request", 10, 1000)
+                    rate_limit("get-request", 10, 1000)
                     # 2000 per 1 minute
-                    rate_limit(conn, "get-request", 60, 2000)
+                    rate_limit("get-request", 60, 2000)
                     # 6000 per 5 minutes
-                    rate_limit(conn, "get-request", 5 * 60, 6000)
+                    rate_limit("get-request", 5 * 60, 6000)
                     # 10000 per 15 minutes
-                    rate_limit(conn, "get-request", 15 * 60, 10000)
+                    rate_limit("get-request", 15 * 60, 10000)
             else:
                 # 10 per 10 seconds
-                rate_limit(conn, "set-request", 10, 10)
+                rate_limit("set-request", 10, 10)
                 # 30 per 1 minute
-                rate_limit(conn, "set-request", 60, 30)
+                rate_limit("set-request", 60, 30)
                 # 100 per 5 minutes
-                rate_limit(conn, "set-request", 5 * 60, 100)
+                rate_limit("set-request", 5 * 60, 100)
                 # 200 per 15 minutes
-                rate_limit(conn, "set-request", 15 * 60, 200)
+                rate_limit("set-request", 15 * 60, 200)
 
 
 # Server health endpoints
