+++ malwarecage/cli/cli.py
@@ -12,8 +12,7 @@ from .database import configure_database
 
 
 @click.group(cls=FlaskGroup, create_app=create_app, add_version_option=False)
-@click.pass_context
-def cli(ctx):
+def cli():
     """Malwarecage malware database"""
