+++ malwarecage/core/service.py
@@ -86,3 +86,6 @@ class Service(Api):
     def relative_url_for(self, resource, **values):
         path = self.url_for(resource, **values)
         return path[len(self.blueprint.url_prefix):]
+
+    def endpoint_for(self, resource):
+        return f'{self.blueprint.name}.{resource}'
